value 값이 같은 경우 || value 값이 더 큰 경우 로 해결해서 query 적음. ItemRepository -> ItemCustomRepository 로 메서드 위치 변경
