+++ src/main/java/com/website/repository/item/customer/ItemCustomerRepository.java
@@ -1,0 +1,7 @@
+package com.website.repository.item.customer;
+
+import com.website.domain.item.Item;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ItemCustomerRepository extends JpaRepository<Item, Long>, ItemCustomerCustomRepository {
+}
