+++ src/main/java/com/website/web/controller/api/home/HomeController.java
@@ -1,8 +1,6 @@
 package com.website.web.controller.api.home;
 
 import com.website.repository.user.UserRepository;
-import com.website.web.dto.request.item.home.ItemHomeSearchCond;
-import com.website.web.dto.sqlcond.item.ItemSearchCond;
 import com.website.web.service.item.ItemService;
 import com.website.web.service.item.carousel.ItemHomeCarouselService;
 import com.website.web.service.item.home.ItemHomeService;
@@ -10,7 +8,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -34,18 +31,19 @@ public class HomeController {
 
     @GetMapping("/main/item/recent")
     public ResponseEntity responseRequestForRecentItem() {
-        return itemHomeService.getItemsReponseLatest();
+        return itemHomeService.getItemsResponseLatest();
     }
 
     @GetMapping("/main/item/special-sale")
     public ResponseEntity responseRequestForSpecialSale() {
-        return itemHomeService.getItemsReponseSpecialSale();
+        return itemHomeService.getItemsResponseSpecialSale();
     }
 
     @GetMapping("/main/item/popular")
     public ResponseEntity responseRequestForPopularItem() {
         return itemHomeService.getItemsResponsePopular();
     }
+
 
     @GetMapping("/item_list")
     public ResponseEntity responseRequestForItemList(
